+++ src/blocks/block-quote.js
@@ -6,10 +6,9 @@ SirTrevor.Blocks.Quote = (function(){
 
   var template = _.template([
     '<blockquote class="st-required st-text-block" contenteditable="true"></blockquote>',
-    '<label class="st-input-label">',
-    i18n.t('blocks:quote:credit_field'),
-    '</label>',
-    '<input maxlength="140" name="cite" placeholder="' + i18n.t('blocks:quote:credit_field') + '" class="st-input-string st-required js-cite-input" type="text" />'
+    '<label class="st-input-label"> <%= i18n.t("blocks:quote:credit_field") %></label>',
+    '<input maxlength="140" name="cite" placeholder="<%= i18n.t("blocks:quote:credit_field") %>"',
+    ' class="st-input-string st-required js-cite-input" type="text" />'
   ].join("\n"));
 
   return SirTrevor.Block.extend({